// careerladder_web/career_ladder/types/jobApplication.ts
export type JobApplication = {
    id: number;
    application_id: string;
    clerk_id: string;
    company_id: string;
    listing_id: string;
    type: string;
    application_status: string;
    project_status: string;
    applied_at: string;
    cover_letter: string;
    skills_fulfilled: string[];
    skills_required: string[];
    resume_url: string;
    resume: string;
    expected_salary: string | null;
    availability: string | null;
    first_name: string;
    last_name: string;
    email: string;
    image_url: string;
    major: string;
    location: string;
    linkedin_url: string;
    profile_summary: string;
    job_preference: string;
    work_status: boolean;
    title: string;
    description: string;
    duration: string;
    allowance: string;
    vacancies: number;
    start_date: string;
    end_date: string;
    created_at: string;
    updated_at: string;
};